Split up the entity relationship tests
