+++ MongoFramework.Tests/EntityRelationships/EntityNavigationCollection/BaseEntityModel.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 
-namespace MongoFramework.Tests.EntityRelationships.EntityCollection
+namespace MongoFramework.Tests.EntityRelationships.EntityNavigationCollection
 {
 	public class BaseEntityModel
 	{
