+++ MongoFramework.Tests/EntityRelationships/Mapping/EntityCollection/InversePropertyModel.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations.Schema;
+using System.Linq;
+
+namespace MongoFramework.Tests.EntityRelationships.Mapping.EntityCollection
+{
+	public class InversePropertyModel
+	{
+		public string Id { get; set; }
+
+		[InverseProperty("RelatedId")]
+		public ICollection<RelatedEntityModel> RelatedEntities { get; set; }
+	}
+}
