+++ MongoFramework.Tests/EntityRelationships/Mapping/SingleEntity/InvalidForeignKeyModel.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.ComponentModel.DataAnnotations.Schema;
+using System.Linq;
+
+namespace MongoFramework.Tests.EntityRelationships.Mapping.SingleEntity
+{
+	public class InvalidForeignKeyModel
+	{
+		public string Id { get; set; }
+
+		[ForeignKey("Created_By")]
+		public string CreatedById { get; set; }
+		public UserEntityModel CreatedBy { get; set; }
+	}
+}
