+++ MongoFramework.Tests/EntityRelationships/Mapping/SingleEntity/UnsupportedIdModel.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.ComponentModel.DataAnnotations.Schema;
+using System.Linq;
+
+namespace MongoFramework.Tests.EntityRelationships.Mapping.SingleEntity
+{
+	public class UnsupportedIdModel
+	{
+		public string Id { get; set; }
+
+		[ForeignKey("CreatedBy")]
+		public int CreatedById { get; set; }
+		public UserEntityModel CreatedBy { get; set; }
+	}
+}
