+++ MongoFramework.Tests/EntityRelationships/Mapping/SingleEntity/UserEntityModel.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Linq;
+
+namespace MongoFramework.Tests.EntityRelationships.Mapping.SingleEntity
+{
+	public class UserEntityModel
+	{
+		public string Id { get; set; }
+		public string Username { get; set; }
+	}
+}
