+++ MongoFramework.Tests/MongoFramework.Tests.csproj
@@ -78,10 +78,20 @@
     <Compile Include="Bson\GetDifferencesTests.cs" />
     <Compile Include="Bson\HasDifferencesTests.cs" />
     <Compile Include="DbEntityCollectionTests.cs" />
-    <Compile Include="EntityRelationships\EntityCollection\BaseEntityModel.cs" />
-    <Compile Include="EntityRelationships\EntityCollection\EntityCollectionTests.cs" />
-    <Compile Include="EntityRelationships\EntityCollection\InversePropertyModel.cs" />
-    <Compile Include="EntityRelationships\EntityCollection\RelatedEntityModel.cs" />
+    <Compile Include="EntityRelationships\EntityNavigationCollection\BaseEntityModel.cs" />
+    <Compile Include="EntityRelationships\EntityNavigationCollection\EntityNavigationCollectionTests.cs" />
+    <Compile Include="EntityRelationships\EntityNavigationCollection\InversePropertyModel.cs" />
+    <Compile Include="EntityRelationships\EntityNavigationCollection\RelatedEntityModel.cs" />
+    <Compile Include="EntityRelationships\Mapping\EntityCollection\BaseEntityModel.cs" />
+    <Compile Include="EntityRelationships\Mapping\EntityCollection\CollectionMappingTests.cs" />
+    <Compile Include="EntityRelationships\Mapping\EntityCollection\InversePropertyModel.cs" />
+    <Compile Include="EntityRelationships\Mapping\EntityCollection\RelatedEntityModel.cs" />
+    <Compile Include="EntityRelationships\Mapping\SingleEntity\BaseEntityModel.cs" />
+    <Compile Include="EntityRelationships\Mapping\SingleEntity\BaseVariedIdModel.cs" />
+    <Compile Include="EntityRelationships\Mapping\SingleEntity\InvalidForeignKeyModel.cs" />
+    <Compile Include="EntityRelationships\Mapping\SingleEntity\SingleEntityMappingTests.cs" />
+    <Compile Include="EntityRelationships\Mapping\SingleEntity\UnsupportedIdModel.cs" />
+    <Compile Include="EntityRelationships\Mapping\SingleEntity\UserEntityModel.cs" />
     <Compile Include="EntityRelationships\Serializer\EntityNavigationCollectionSerializerTests.cs" />
     <Compile Include="EntityRelationships\SingleEntity\InvalidForeignKeyModel.cs" />
     <Compile Include="EntityRelationships\SingleEntity\UnsupportedIdModel.cs" />
